Move set_name into device
