+++ src/core/traits.rs
@@ -1,9 +1,6 @@
 //! Defines traits for core
 
-use anyhow::Result;
 use ash::vk;
-
-use crate::Device;
 
 /// Direct access to the object's handle's as_raw representation
 pub unsafe trait AsRaw {
@@ -15,22 +12,4 @@ pub unsafe trait AsRaw {
 pub trait Nameable: AsRaw {
     /// Change the name of the given object
     const OBJECT_TYPE: vk::ObjectType;
-}
-
-impl Device {
-    /// Set the name of any given compatible object for debugging purposes
-    pub fn set_name<T: Nameable>(&self, object: &T, name: &str) -> Result<()> {
-        let object_name = std::ffi::CString::new(name)?;
-        let name_info = vk::DebugUtilsObjectNameInfoEXT::builder()
-            .object_type(<T as Nameable>::OBJECT_TYPE)
-            .object_handle(unsafe { object.as_raw() })
-            .object_name(&object_name)
-            .build();
-
-        unsafe {
-            Ok(self
-                .debug_utils()?
-                .set_debug_utils_object_name(self.handle().handle(), &name_info)?)
-        }
-    }
 }
